test: split e2e specs and tag smoke coverage
